Remove ResourceManager and Resource classes that are no longer needed

better label selectors


remove ResourceManager completely


create CrdResourceClients instead of keeping the clients in each utils class)


fixups


Fixups


comments from Jakub, Kali, and Maros


revert docs stuff


javadoc and some comments


Rename KafkaResourceNames to KafkaComponents


changes after rebase and one more fix

+++ systemtest/src/main/java/io/strimzi/systemtest/resources/operator/SetupClusterOperator.java
@@ -118,8 +118,13 @@ public class SetupClusterOperator {
      */
     public void createClusterOperatorNamespace() {
         if (KubeResourceManager.get().kubeClient().namespaceExists(clusterOperatorConfiguration.getNamespaceName())) {
-            Namespace coNamespace = KubeResourceManager.get().kubeClient().getClient().namespaces().withName(clusterOperatorConfiguration.getNamespaceName()).get();
-            KubeResourceManager.get().deleteResource(coNamespace);
+            if (clusterOperatorConfiguration.shouldDeleteNamespace()) {
+                Namespace coNamespace = KubeResourceManager.get().kubeClient().getClient().namespaces().withName(clusterOperatorConfiguration.getNamespaceName()).get();
+                KubeResourceManager.get().deleteResource(coNamespace);
+            } else {
+                // in case we are skipping the deletion, there is no need recreating the Namespace
+                return;
+            }
         }
 
         KubeResourceManager.get().createResourceWithWait(new NamespaceBuilder()
